+++ PKHeX.WinForms/Subforms/Save Editors/Gen7/SAV_HallOfFame7.cs
@@ -61,7 +61,7 @@ namespace PKHeX.WinForms
             int ofs = SAV.HoF;
 
             SAV.Data[ofs] &= 0xFE;
-            SAV.Data[ofs] |= (byte)(CHK_Flag.Checked ? 1 : 0);
+            SAV.Data[ofs] |= CHK_Flag.Checked ? 1 : 0;
             BitConverter.GetBytes((ushort)NUD_Count.Value).CopyTo(SAV.Data, ofs + 2);
             for (int i = 0; i < entries.Length; i++)
             {
